Encode: Split builder pattern
Separate structs for different stages of the encoding make the code more
readable. The Encoder handles the low-level operations, which Program,
Witness and IllegalPadding call. This also means that Encoder no longer
needs to follow the builder pattern. Shenanigans and abuse of the
encoding are still possible, but the intend behind the encoded program
can be better conveyed with readable method names.